# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Fieldnotes Blog API
  version: 1.0.0
  description: 블로그 사용자/관리자 API v1
servers:
  - url: http://{SERVER_URL}:8080/
    variables:
      SERVER_URL:
        default: localhost
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── 공통 응답 ──────────────────────────────
  responses:
...
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 404
            code: "E404_001"
            detail: "resource not found."
            message: "요청한 리소스를 찾을 수 없습니다."
            timestamp: "2025-12-26T01:34:00+09:00"
            path: "/api/v1/..."
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 400
            code: "E400_002"
            detail: "'AccountID' does not satisfy 'required' condition"
            message: "요청 파라미터의 타입 또는 형식이 올바르지 않습니다."
            timestamp: "2026-05-05T11:40:38.485176711Z"
            path: "/api/v1/..."
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 401
            code: "E401_001"
            detail: "Invalid admin credentials."
            message: "인증에 실패하였습니다."
            timestamp: "2025-12-26T01:34:00+09:00"
            path: "/api/v1/..."
...
    InternalServerError:
      description: 서버 내부 오류
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: 500
            code: "E500_001"
            detail: "failed to encode args[0]..."
            message: "서버 내부 에러가 발생하였습니다."
            timestamp: "2026-05-06T02:57:20.333687031Z"
            path: "/api/v1/..."

  # ── 재사용 파라미터 ───────────────────────────
  parameters:
...
    PathAccount:
      name: accountId
      in: path
      required: true
      schema:
        type: string
        default: testAccount1
      description: 관리자 계정 ID
...
  schemas:
...
    CommonResponse:
      type: object
      required: [status, code, detail, message, timestamp, path]
      properties:
        status:
          type: integer
          example: 200
        code:
          type: string
          example: "E200_001"
        detail:
          type: string
          example: "Successfully retrieved."
        message:
          type: string
          example: "정상적으로 처리되었습니다."
        timestamp:
          type: string
          format: date-time
          example: "2026-01-20T16:31:00+09:00"
        path:
          type: string
          example: "/api/v1/posts"
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/CommonResponse'
...
    AccountDetail:
      type: object
      required: [id, accountId, nickname, avatarUrl, role, status, createdAt, updatedAt]
      properties:
        id:
          type: integer
          example: 5
        accountId:
          type: string
          example: "admin"
        nickname:
          type: string
          example: "admin"
        avatarUrl:
          type: string
          example: "https://www.svgrepo.com/show/345423/admin.svg"
        role:
          type: string
          example: "ADMIN"
        status:
          type: string
          example: "ACTIVE"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /api/v1/auth/{accountId}:
    get:
      tags: [Admin]
      summary: 관리자 계정 조회
      operationId: getAccount
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathAccount'
      responses:
        '200':
          description: 계정 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/CommonResponse'
                  - type: object
                    properties:
                      result:
                        $ref: '#/components/schemas/AccountDetail'
              example:
                status: 200
                code: "E200_001"
                detail: "Successfully retrieved 'admin' account."
                message: "정상적으로 처리되었습니다."
                timestamp: "2026-05-12T04:30:26.826058992Z"
                path: "/api/v1/auth/admin"
                result:
                  id: 4
                  accountId: "admin"
                  nickname: "admin"
                  avatarUrl: "https://www.svgrepo.com/show/345423/admin.svg"
                  role: "ADMIN"
                  status: "ACTIVE"
                  createdAt: "2026-05-12T13:27:51.404057+09:00"
                  updatedAt: "2026-05-12T13:27:51.404057+09:00"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...